+++ src/prepare_rgz_data.py
@@ -63,13 +63,23 @@ def main():
     rgz_path = os.path.join(cwd, 'data/rgz')
     download_path = os.path.join(rgz_path, 'download/')
     collect_path = os.path.join(rgz_path, 'csv/')
+    all_addresses_path = os.path.join(rgz_path, 'addresses.csv')
+    if os.path.exists(all_addresses_path):
+        need_address_collection = False
+
     all_addresses = []
     total_zips = len(os.listdir(download_path))
     for i, file in enumerate(os.listdir(download_path)):
         if not file.endswith(".zip"):
             continue
         opstina = file[:-4]
-        print(f"{i+1}/{total_zips} Processing {opstina}")
+        opstina_csv_path = os.path.join(collect_path, opstina + ".csv")
+
+        print(f"{i+1}/{total_zips} Processing {opstina}... ", end='')
+        if os.path.exists(opstina_csv_path) and not need_address_collection:
+            print(f"skipping, file data/rgz/csv/{opstina}.csv already exists")
+            continue
+
         with zipfile.ZipFile(os.path.join(download_path, file), 'r') as zip_ref:
             zip_ref.extract(opstina + ".csv", rgz_path)
         temp_opstina_csv_file = os.path.join(rgz_path, opstina + ".csv")
@@ -77,7 +87,11 @@ def main():
         all_addresses += opstina_addresses
         os.remove(temp_opstina_csv_file)
 
-        with open(os.path.join(collect_path, opstina + ".csv"), 'w') as opstina_addresses_csv:
+        if os.path.exists(opstina_csv_path):
+            print(f"skipping, file data/rgz/csv/{opstina}.csv already exists")
+            continue
+
+        with open(opstina_csv_path, 'w') as opstina_addresses_csv:
             writer = csv.DictWriter(
                 opstina_addresses_csv, fieldnames=['rgz_kucni_broj_id', 'rgz_opstina_mb', 'rgz_opstina',
                                                'rgz_naselje_mb', 'rgz_naselje', 'rgz_ulica_mb', 'rgz_ulica',
@@ -85,9 +99,14 @@ def main():
             writer.writeheader()
             for address in opstina_addresses:
                 writer.writerow(address)
+        print("OK")
 
-    print("Collecting all addresses")
-    all_addresses_path = os.path.join(rgz_path, 'addresses.csv')
+    print("Collecting all addresses...", end='')
+
+    if os.path.exists(all_addresses_path):
+        print("skipping, already exists")
+        return
+
     with open(all_addresses_path, 'w') as all_addresses_csv:
         writer = csv.DictWriter(
             all_addresses_csv, fieldnames=['rgz_kucni_broj_id', 'rgz_opstina_mb', 'rgz_opstina',
@@ -96,7 +115,7 @@ def main():
         writer.writeheader()
         for address in all_addresses:
             writer.writerow(address)
-    print(f"All {len(all_addresses)} addresses written to data/rgz/addresses.csv")
+    print(f"OK ({len(all_addresses)} addresses written)")
 
 
 if __name__ == '__main__':
